Some more event edits for Byzantium
War with Carthage, cores on Kachuskaya, more atrocities yaaay

Soon will be wars with Aquileia and the Pope. The first will be in response to border disputes in Italy, and the latter as a response to the Byzantine excommunication. After all conquests are complete, Byzantium will look to centralize its rule in Italy. This will be more or less the end of Byzantine content, save for special events recognizing Byzantine accomplishments elsewhere and/or Armenia/Caucasus expansion. Byzantium WILL NOT accept italian pops, and will generally remain a greek-only state